+++ src/NLGAMap/NLGAMap.js
@@ -11,6 +11,7 @@ import {SVGRectPattern} from '../Utils/SVGRectPattern';
 
 import {Layers} from './Layers/Layers';
 import {MarkerMap} from './Marker/MarkerMap/MarkerMap';
+import {PieMap} from './Marker/MarkerMap/MarkerMap.Pie';
 import {SymbolArrows} from './Marker/SymbolMap/Arrows/SymbolArrows';
 import {SymbolBars} from './Marker/SymbolMap/Bars/SymbolBars';
 import {Choropleth} from './Choropleth/Choropleth';
@@ -34,6 +35,7 @@ export default class NLGAMap {
         this.choropleths = {};
         this.markerMaps  = {};
         this.symbolMaps  = {};
+        this.tileMaps    = {};
     }
 
     static get VERSION() {
@@ -71,12 +73,17 @@ export default class NLGAMap {
     }
 
     addMarkerMap(options) {
-        let markerMap = new MarkerMap(options, this.layers),
+        let markerMap = (options.type === 'pie') ? new PieMap(options, this.layers) : new MarkerMap(options, this.layers),
         promise       = markerMap.addLegendTo(this.leaflet).render();
 
         this.markerMaps[options.layerName] = markerMap;
 
         return promise;
+    }
+
+    addTileMap(options) {
+        let tileMap = this.layers.addTileLayer(options);
+        this.tileMaps[options.layerName] = tileMap;
     }
 
     addChoroplethMap(options) {
@@ -142,14 +149,22 @@ export default class NLGAMap {
 
     destroy() {
         this.layers.destroy();
+        each(this.tileMaps, (tileMap) => {tileMap.destroy();});
         each(this.choropleths, (choropleth) => {choropleth.destroy();});
         each(this.markerMaps, (markerMap) => {markerMap.destroy();});
         each(this.symbolMaps, (symbolMap) => {symbolMap.destroy();});
+
+        if (this._zoomControl) {
+            this._zoomControl.remove();
+            this._zoomControl = null;
+        }
+
         this.leaflet.remove();
 
         this.options     = null;
         this.container   = null;
         this.layers      = null;
+        this.tileMaps    = null;
         this.choropleths = null;
         this.markerMaps  = null;
         this.symbolMaps  = null;
@@ -160,11 +175,14 @@ export default class NLGAMap {
         if (this.leaflet) this.leaflet.remove();
         this.leaflet = L.map(this.id, mapOptions);
         
-        if (mapOptions.controls.zoom)
-            L.control.zoom(mapOptions.controls.zoom).addTo(this.leaflet);
+        if (mapOptions.controls.zoom) {
+            this._zoomControl = L.control.zoom(mapOptions.controls.zoom).addTo(this.leaflet);
+        }
 
-        if (this.options.map.fullscreenControl === true)
-            this._addFullscreenChangeListener();
+        if (this.options.map.fullscreenControl === true) {
+            this._onFullscreenChange = this._onFullscreenChange.bind(this);
+            this.leaflet.on('fullscreenchange', this._onFullscreenChange);
+        }
 
         if (L.Browser.mobile) 
             this.leaflet.dragging.disable();
@@ -186,10 +204,11 @@ export default class NLGAMap {
     }
 
     _addUserLayers() {
-        let promises = [],
-        choropleth   = this.options.layers.choropleth,
-        markers      = this.options.layers.markers,
-        symbols      = this.options.layers.symbols;
+        let promises   = [],
+            choropleth = this.options.layers.choropleth,
+            markers    = this.options.layers.markers,
+            symbols    = this.options.layers.symbols,
+            tiles      = this.options.layers.tiles;
 
         if (choropleth) {
             choropleth.map_id       = this.id;
@@ -214,35 +233,43 @@ export default class NLGAMap {
             });
         }
 
+        if (tiles) {
+            tiles.forEach((tileMap) => {
+                tileMap.map_id = this.id;
+                this.addTileMap(tileMap);
+            });
+        }
+
         return promises;
     }
 
-    _addFullscreenChangeListener() {
-        this.leaflet.on('fullscreenchange', () => {
-            if (this.leaflet.isFullscreen()) {
-                const bounds                   = this.layers[this.options.baseLayer.layerName].getBounds();
-                      this._current_zoom_level = this.leaflet.getZoom(),
-                      this._current_center     = this.leaflet.getCenter();
-                
-                this.leaflet.setMaxBounds(bounds.pad(0.5));
-                this.leaflet.setMaxZoom(10);
-                this.leaflet.fitBounds(bounds);
+    _onFullscreenChange() {
+        if (this.leaflet.isFullscreen()) {
+            const bounds             = this.layers[this.options.baseLayer.layerName].getBounds();
+            this._current_zoom_level = this.leaflet.getZoom();
+            this._current_center     = this.leaflet.getCenter();
+            
+            this.leaflet.setMaxBounds(bounds.pad(0.5));
+            this.leaflet.setMaxZoom(10);
+            this.leaflet.fitBounds(bounds);
 
-                if (L.Browser.mobile)
-                    this.leaflet.dragging.enable();
-            } else {
-                this.leaflet.setMaxBounds(this.options.map.maxBounds);
-                this.leaflet.setMaxZoom(this.options.map.maxZoom);
-                
-                this.leaflet.setView(this._current_center, this._current_zoom_level, {animate: false});
-
-                if (this.options.map.fitBounds === true)
-                    this.leaflet.fitBounds(this.layers[this.options.baseLayer.layerName].getBounds());
-
-                if (L.Browser.mobile)
-                    this.leaflet.dragging.disable();
+            if (L.Browser.mobile) {
+                this.leaflet.dragging.enable();
             }
-        });
+        } else {
+            this.leaflet.setMaxBounds(this.options.map.maxBounds);
+            this.leaflet.setMaxZoom(this.options.map.maxZoom);
+            
+            this.leaflet.setView(this._current_center, this._current_zoom_level, {animate: false});
+
+            if (this.options.map.fitBounds === true) {
+                this.leaflet.fitBounds(this.layers[this.options.baseLayer.layerName].getBounds());
+            }
+
+            if (L.Browser.mobile) {
+                this.leaflet.dragging.disable();
+            }
+        }
     }
 
     _addDefs(patterns, colors) {